+++ cmd/scan.go
@@ -12,6 +12,7 @@ import (
 	"github.com/malandas/andas/internal/scanner"
 	"github.com/malandas/andas/internal/scanner/deps"
 	"github.com/malandas/andas/internal/scanner/githistory"
+	"github.com/malandas/andas/internal/scanner/sast"
 	"github.com/malandas/andas/internal/scanner/secrets"
 )
 
@@ -77,6 +78,7 @@ func runScan(args []string) int {
 	scanners := []scanner.Scanner{
 		secrets.New(),
 		deps.New(),
+		sast.New(),
 	}
 	if *history {
 		scanners = append(scanners, githistory.New())